i2c poll: fix communication with multiple slaves
for some reason polling right when pushing a request to the interface
will stall communication, always going into the error handler.

calling `poll` regularly in its own 'task' instead seems to fix this
particular bug.

this probably only fixes the symptoms instead of the actual reason
behind it, but works for now. more investigation will be necessary to
find the underlying reason and find a proper fix.