kern: add exit functionality
this is necessary to enable code to actually clean up properly after
itself, as calling the OS' `exit()` does not unwind the stack and hence
does not call destructors.
This has not been a problem until now, because we weren't using
resources which the OS doesn't clean up after process termination
anyway. Now that we start using hardware through the `sysfs` this
changes.

This will also enable us to write short testing programs that rely on
the kernel running, without going through the debugger for cleanup.

note that this functionality should probably _not_ be used on the STM
backend, as there be dragons when `main()` exits.