introduce sys/comm.h on linux
this implements UDP and TTY communication _without_ relying on `ncat`
and `poll` -- turns out nonblocking communication can be dealt with more
easily, and we've already been doing it for CAN.

also introduces an IFACE (read: interface) on top of TTY & Logger for
e.g. writing simple interactive user interfaces.

CAN, UDP, and TTY all have basically the same structure and
implementation and could probably be unified, that remains as a TODO for
later if it turns out to be helpful. Didn't manage to wrap my head
around it as of yet.