remove dependency on `ncat`, deprecate sys/host.h
removes special treatment of the linux backend -- not all programs need
both the UDP and TTY interfaces -- they shall declare and use only what
they need.

also gets rid of the `ncat` dependency, _and_ -- equally important --
the need to run in multiple threads. This still did not allow cleaning
up the program gracefully and running all destructors.

The UDP implementation arguably works better than communicating with
`ncat` -- no need to wait for `ncat` to restart after the 'connection
drops' -- UDP is connectionless, this implementation reflects that.

`sys/host.h` is still available for now, implementing the old `host`
struct on top of the new backends, but I'm in favor of removing it
completely.