BNO085 (#123)
* Start with BNO085 as i2c device

* Add link to sh reference manual for bno

* uart: make sure to have reserved at least one ms

* i2c: be more liberal in what we expect to happen

i.e. don't assert things we're not actually sure about

* logger: add Buffered Logger

* buffer: support copying arbitrary data types

this previously only worked on Plain Old Data objects

* queue: support copying and moving queues

* schedule: compiler happiness

* core/timed: check dt_ms before using recurring calls

* add utils/slice.h

this is a way to pass around views of data without having to copy or
allocate it.

TODO: consider making this a view of raw data instead of Buffers

* dev/BNO085: raw & unpolished, but first features work

* dev/BNO085: cleanup implementation, documentation

* CI: add Debug builds

* utils/later: fix leak

this seems to have gone unnoticed for the longest time

* queue: simplify & fix

disallow queues of types that are not default-constructible.
working around this limitation doesn't actually seem to make sense

* bno085: documentation work

---------

Co-authored-by: Jonathan Halmen <slowjo@halmen.xyz>